// src/libs/pb/readly/v1/rpc_sign_in_pb.d.ts
// @generated by protoc-gen-es v2.2.5
// @generated from file readly/v1/rpc_sign_in.proto (package readly.v1, syntax proto3)
/* eslint-disable */

import type { GenFile, GenMessage } from "@bufbuild/protobuf/codegenv1";
import type { Message } from "@bufbuild/protobuf";

/**
 * Describes the file readly/v1/rpc_sign_in.proto.
 */
export declare const file_readly_v1_rpc_sign_in: GenFile;

/**
 * @generated from message readly.v1.SignInRequest
 */
export declare type SignInRequest = Message<"readly.v1.SignInRequest"> & {
  /**
   * @generated from field: string email = 1;
   */
  email: string;

  /**
   * @generated from field: string password = 2;
   */
  password: string;
};

/**
 * Describes the message readly.v1.SignInRequest.
 * Use `create(SignInRequestSchema)` to create a new message.
 */
export declare const SignInRequestSchema: GenMessage<SignInRequest>;

/**
 * @generated from message readly.v1.SignInResponse
 */
export declare type SignInResponse = Message<"readly.v1.SignInResponse"> & {
  /**
   * @generated from field: string access_token = 1;
   */
  accessToken: string;

  /**
   * @generated from field: string refresh_token = 2;
   */
  refreshToken: string;

  /**
   * @generated from field: int64 user_id = 3;
   */
  userId: bigint;

  /**
   * @generated from field: string name = 4;
   */
  name: string;

  /**
   * @generated from field: string email = 5;
   */
  email: string;
};

/**
 * Describes the message readly.v1.SignInResponse.
 * Use `create(SignInResponseSchema)` to create a new message.
 */
export declare const SignInResponseSchema: GenMessage<SignInResponse>;