// src/libs/pb/readly/v1/rpc_sign_up_pb.d.ts
// @generated by protoc-gen-es v2.2.5
// @generated from file readly/v1/rpc_sign_up.proto (package readly.v1, syntax proto3)
/* eslint-disable */

import type { GenFile, GenMessage } from "@bufbuild/protobuf/codegenv1";
import type { Message } from "@bufbuild/protobuf";

/**
 * Describes the file readly/v1/rpc_sign_up.proto.
 */
export declare const file_readly_v1_rpc_sign_up: GenFile;

/**
 * @generated from message readly.v1.SignUpRequest
 */
export declare type SignUpRequest = Message<"readly.v1.SignUpRequest"> & {
  /**
   * @generated from field: string name = 1;
   */
  name: string;

  /**
   * @generated from field: string email = 2;
   */
  email: string;

  /**
   * @generated from field: string password = 3;
   */
  password: string;
};

/**
 * Describes the message readly.v1.SignUpRequest.
 * Use `create(SignUpRequestSchema)` to create a new message.
 */
export declare const SignUpRequestSchema: GenMessage<SignUpRequest>;

/**
 * @generated from message readly.v1.SignUpResponse
 */
export declare type SignUpResponse = Message<"readly.v1.SignUpResponse"> & {
  /**
   * @generated from field: string access_token = 1;
   */
  accessToken: string;

  /**
   * @generated from field: string refresh_token = 2;
   */
  refreshToken: string;

  /**
   * @generated from field: int64 user_id = 3;
   */
  userId: bigint;

  /**
   * @generated from field: string name = 4;
   */
  name: string;

  /**
   * @generated from field: string email = 5;
   */
  email: string;
};

/**
 * Describes the message readly.v1.SignUpResponse.
 * Use `create(SignUpResponseSchema)` to create a new message.
 */
export declare const SignUpResponseSchema: GenMessage<SignUpResponse>;